moduler listener design
