+++ index.js
@@ -3,11 +3,15 @@
 const fs = require('fs');
 const Discord = require('discord.js');
 const { prefix, token } = require('./config.json');
+const fg = require("./util/FileGetter.js")
 //Preload Command Registry
 const client = new Discord.Client();
-client.commands = new Discord.Collection();
+client.prefix = prefix; //laziness here
+client.commands = new Discord.Collection(); //perhaps extend client instead
+client.cooldowns = new Discord.Collection();
 
-const commandFiles = require("./util/FileGetter.js").getFiles("./commands");
+const commandFiles = fg.getFiles("./commands");
+const listenerFiles = fg.getFiles("./listeners")
 
 for (const file of commandFiles) {
 	const command = require(`${file}`);
@@ -15,7 +19,12 @@ for (const file of commandFiles) {
     console.log(`Command ${file} loaded!`);
 }
 
-const cooldowns = new Discord.Collection();
+for (const file of listenerFiles) {
+	const listener = require(`${file}`);
+	const name = file.replace("./listeners/", '').split(".")[0]; //todo: regex
+	console.log(`Listening on ${name}...`);
+	client.on(name, (...args) => listener(client, ...args));
+}
 
 client.login(token);
 
@@ -24,57 +33,4 @@ client.once('ready', () => {
     client.user.setActivity("the Imposters :eyes:", {
         type: "WATCHING"
     });
-});
-//Command Handler
-client.on('message', message => {
-	if (!message.content.startsWith(prefix) || message.author.bot) return;
-
-	const args = message.content.slice(prefix.length).trim().split(/ +/);
-	const commandName = args.shift().toLowerCase();
-
-	const command = client.commands.get(commandName)
-		|| client.commands.find(cmd => cmd.aliases && cmd.aliases.includes(commandName));
-
-	if (!command) return;
-
-	if (command.guildOnly && message.channel.type === 'dm') {
-		return message.reply('I can\'t execute that command inside DMs!');
-	}
-
-	if (command.args && !args.length) {
-		let reply = `You didn't provide any arguments, ${message.author}!`;
-
-		if (command.usage) {
-			reply += `\nThe proper usage would be: \`${prefix}${command.name} ${command.usage}\``;
-		}
-
-		return message.channel.send(reply);
-	}
-
-	if (!cooldowns.has(command.name)) {
-		cooldowns.set(command.name, new Discord.Collection());
-	}
-
-	const now = Date.now();
-	const timestamps = cooldowns.get(command.name);
-	const cooldownAmount = (command.cooldown || 3) * 1000;
-
-	if (timestamps.has(message.author.id)) {
-		const expirationTime = timestamps.get(message.author.id) + cooldownAmount;
-
-		if (now < expirationTime) {
-			const timeLeft = (expirationTime - now) / 1000;
-			return message.reply(`please wait ${timeLeft.toFixed(1)} more second(s) before reusing the \`${command.name}\` command.`);
-		}
-	}
-
-	timestamps.set(message.author.id, now);
-	setTimeout(() => timestamps.delete(message.author.id), cooldownAmount);
-
-	try {
-		command.execute(client, message, args);
-	} catch (error) {
-		console.error(error);
-		message.reply('there was an error trying to execute that command!');
-	}
 });
